coredump: save process container parent cmdline
Process container parent is the process used to start processes with a new
user namespace - e.g systemd-nspawn, runc, lxc, etc.

There is not standard way how to find that process - or I do not
know about it, so I decided to find the first process in the parent
process hierarchy using a different mount namespace.

This information is at least useful for bug reporting tools like ABRT.